requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
Django==4.0.3
django-spa==0.3.6
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
sqlparse==0.4.2
tomli==2.0.1
whitenoise==5.3.0
request==